--- app/globals.css
@@ -1,16 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-* {
-  user-select: none;
-}
-
-@layer components {
-  .page {
-    @apply w-full flex-grow flex flex-col mt-14 max-w-[40rem] px-8 md:w-[40rem] py-12;
-  }
-  .page-wide {
-    @apply w-3/4 flex-grow flex flex-col mt-14 px-8 py-12;
-  }
-}
