--- app/projects/page.tsx
@@ -1,7 +1,0 @@
-export default function ProjectPage() {
-  return (
-    <main className="page">
-      <p>Hellow world from dev with kd</p>
-    </main>
-  );
-}
