--- app/resume/page.tsx
@@ -1,7 +1,0 @@
-export default function Home() {
-  return (
-    <main className="page">
-      <p>Hellow world from dev with kd</p>
-    </main>
-  );
-}
